% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nvd-class.R
\name{add_null_nodes}
\alias{add_null_nodes}
\title{Adding null nodes to networks}
\usage{
add_null_nodes(x, num_vertices = NULL, directed = FALSE, weighted = FALSE)
}
\arguments{
\item{x}{An \code{nvd} object.}

\item{num_vertices}{An integer specifying the desired number of vertices of the networks. If necessary, null nodes are added in each network to reach this number.  If \code{NULL}, the maximum number of vertices in the sample is considered. Default to \code{NULL}.}

\item{directed}{A boolean specifying whether the function should consider the input networks as directed or undirected. Default to \code{FALSE}.}

\item{weighted}{A boolean specifying whether the function should consider the input networks as weighted or unweighted Default to \code{FALSE}.}
}
\value{
An \code{nvd} object.
}
\description{
This function adds null nodes to networks stored in an \code{nvd} object.
}
\examples{
gnp_params <- list(p = 1/3)
x <- nvd(model = "gnp", n = 10L, num_vertices = 10L, model_params = gnp_params)
add_null_nodes(x, num_vertices = 12L, directed = FALSE, weighted = FALSE)
}
