% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nvd-class.R
\name{as_nvd}
\alias{as_nvd}
\title{Coercion to Network-Valued Data Object}
\usage{
as_nvd(obj)
}
\arguments{
\item{obj}{A list of \code{\link[igraph]{igraph}} objects.}
}
\value{
An \code{\link{nvd}} object.
}
\description{
This function flags a list of \code{\link[igraph]{igraph}} objects as an
\code{\link{nvd}} object as defined in this package.
}
\examples{
gnp_params <- list(p = 1/3)
as_nvd(nvd(model = "gnp", n = 10L, model_params = gnp_params))
}
