% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertex_partition.R
\name{as_vertex_partition}
\alias{as_vertex_partition}
\title{Coerion to Vertex Partition}
\usage{
as_vertex_partition(x)
}
\arguments{
\item{x}{A list grouping the vertices by partition element or an integer or
character vector of vertex memberships.}
}
\value{
A \code{vertex_partition} object storing the corresponding vertex
partition.
}
\description{
This function converts a vector of memberships into a proper vertex partition
object.
}
\examples{
m1 <- c("P1", "P3", "P4", "P1", "P2", "P2", "P3", "P1", "P4", "P3")
V1 <- as_vertex_partition(m1)
m2 <- as.integer(c(1, 3, 4, 1, 2, 2, 3, 1, 4, 3))
V2 <- as_vertex_partition(m2)
}
