% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nvd-class.R
\name{bind_nvd}
\alias{bind_nvd}
\title{Concatenation of \code{nvd} objects}
\usage{
bind_nvd(...)
}
\arguments{
\item{...}{A number of \code{\link{nvd}} objects.}
}
\value{
A network-valued data set as \code{\link{nvd}} object.
}
\description{
Concatenation of \code{nvd} objects
}
\examples{
x <- nvd("pa", 5)
y <- nvd("smallworld", 5)
z <- bind_nvd(x, y)
}
