% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_distances.R
\name{get-distance}
\alias{get-distance}
\alias{get_hamming_distance}
\alias{get_frobenius_distance}
\alias{get_spectral_distance}
\alias{get_root_euclidean_distance}
\title{Distances Between Networks}
\usage{
get_hamming_distance(x, y, representation = "laplacian")

get_frobenius_distance(x, y, representation = "laplacian")

get_spectral_distance(x, y, representation = "laplacian")

get_root_euclidean_distance(x, y, representation = "laplacian")
}
\arguments{
\item{x}{An \code{\link[igraph]{igraph}} object or a matrix representing an
underlying network.}

\item{y}{An \code{\link[igraph]{igraph}} object or a matrix representing an
underlying network. Should have the same number of vertices as \code{x}.}

\item{representation}{A string specifying the desired type of representation,
among: \code{"adjacency"}, \code{"laplacian"} [default] and
\code{"modularity"}.}
}
\value{
A scalar measuring the distance between the two input networks.
}
\description{
This is a collection of functions computing the distance between two
networks.
}
\details{
Let \eqn{X} be the matrix representation of network \eqn{x} and \eqn{Y} be
the matrix representation of network \eqn{y}. The Hamming distance between
\eqn{x} and \eqn{y} is given by \deqn{\frac{1}{N(N-1)} \sum_{i,j} |X_{ij} -
Y_{ij}|,} where \eqn{N} is the number of vertices in networks \eqn{x} and
\eqn{y}. The Frobenius distance between \eqn{x} and \eqn{y} is given by
\deqn{\sqrt{\sum_{i,j} (X_{ij} - Y_{ij})^2}.} The spectral distance between
\eqn{x} and \eqn{y} is given by \deqn{\sqrt{\sum_i (a_i - b_i)^2},} where
\eqn{a} and \eqn{b} of the eigenvalues of \eqn{X} and \eqn{Y}, respectively.
This distance gives rise to classes of equivalence. Consider the spectral
decomposition of \eqn{X} and \eqn{Y}: \deqn{X=VAV^{-1}} and \deqn{Y =
UBU^{-1},} where \eqn{V} and \eqn{U} are the matrices whose colums are the
eigenvectors of \eqn{X} and \eqn{Y}, respectively and \eqn{A} and \eqn{B} are
the diagonal matrices with elements the eigenvalues of \eqn{X} and \eqn{Y},
respectively. The root-Euclidean distance between \eqn{x} and \eqn{y} is
given by \deqn{\sqrt{\sum_i (V \sqrt{A} V^{-1} - U \sqrt{B} U^{-1})^2}.}
Root-Euclidean distance can used only with the laplacian matrix
representation.
}
\examples{
g1 <- igraph::erdos.renyi.game(20, 0.1)
g2 <- igraph::erdos.renyi.game(20, 0.2)
get_hamming_distance(g1, g2, "adjacency")
get_frobenius_distance(g1, g2, "adjacency")
get_spectral_distance(g1, g2, "laplacian")
get_root_euclidean_distance(g1, g2, "laplacian")
}
