% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_representations.R
\name{get-representation}
\alias{get-representation}
\alias{get_adjacency}
\alias{get_laplacian}
\alias{get_modularity}
\title{Network Representation Functions}
\usage{
get_adjacency(network, validate = TRUE)

get_laplacian(network, validate = TRUE)

get_modularity(network, validate = TRUE)
}
\arguments{
\item{network}{An \code{\link[igraph]{igraph}} object.}

\item{validate}{A boolean specifying whether the function should check the
class of its input (default: \code{TRUE}).}
}
\value{
A numeric square matrix giving the desired network representation
  recorded in the object's class.
}
\description{
This is a collection of functions that convert a graph stored as an
\code{\link[igraph]{igraph}} object into a desired matrix representation
among adjacency matrix, graph laplacian or modularity matrix.
}
\examples{
X <- igraph::watts.strogatz.game(1, 25, 3, 0.05)
adjX <- get_adjacency(X)
lapX <- get_laplacian(X)
modX <- get_modularity(X)
}
