% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_statistics.R
\name{get-statistic}
\alias{get-statistic}
\alias{get_average_statistic}
\alias{get_frechet_statistic}
\title{Test Statistics for Network Populations}
\usage{
get_average_statistic(d, indices)

get_frechet_statistic(d, indices)
}
\arguments{
\item{d}{A matrix of dimension \eqn{(n1+n2)x(n1+n2)} containing the
distances between all the elements of the two samples put together.}

\item{indices}{A vector of dimension \eqn{n1} containing the indices of the
elements of the first sample.}
}
\value{
A scalar giving the value of the desired test statistic.
}
\description{
This is a collection of functions that provide various test statistics for
populations of networks. The \emph{average} test statistic is given by the
mean of all the distances between the \eqn{n1} elements of the fist sample
and those (\eqn{n2}) of the second one. The \emph{Fréchet} test statistic is
given by the distance between the Frechet mean of the first sample (with
cardinality \eqn{n1}) and that of the second one (with cardinality \eqn{n2}).
The distance used to compute the value of the test statistic is the same used
to find the Fréchet means.
}
