% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nvd-class.R
\name{is_nvd}
\alias{is_nvd}
\title{Test nvd}
\usage{
is_nvd(obj)
}
\arguments{
\item{obj}{An R object.}
}
\value{
A boolean specifying whether the input R object is of class
\code{\link{nvd}}.
}
\description{
Test nvd
}
