% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nevada.R
\docType{package}
\name{nevada-package}
\alias{nevada}
\alias{nevada-package}
\title{nevada: network-valued data analysis}
\description{
nevada provides a ...
}
\details{
It has three main goals:

\itemize{
\item ...
\item ...
\item ...
}

To learn more about nevada, start with the vignettes:
`browseVignettes(package = "nevada")`
}
