% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{nvd-plot}
\alias{nvd-plot}
\alias{autoplot.nvd}
\alias{plot.nvd}
\title{MDS Visualization of Network Distributions}
\usage{
\method{autoplot}{nvd}(x, y, representation = NULL, distance = NULL, ...)

\method{plot}{nvd}(x, y, representation = NULL, distance = NULL, ...)
}
\arguments{
\item{x}{A \code{\link{nvd}} object.}

\item{y}{A \code{\link{nvd}} object.}

\item{representation}{A character vector specifying the chosen representation(s),
among: \code{"adjacency"}, \code{"laplacian"},
\code{"modularity"}. If \code{NULL}, each representation is chosen. Defaults to \code{NULL}.}

\item{distance}{A character vector specifying the chosen distance(s),
among: \code{"frobenius"}, \code{"hamming"},
\code{"spectral"}, \code{"root-euclidean"} and \code{"match-frobenius"}. If \code{NULL}, each distance is chosen. Defaults to \code{NULL}.}

\item{...}{Extra arguments to be passed to the plot function.}
}
\value{
Invisibly returns a \code{\link[ggplot2]{ggplot}} object. In
particular, the data set computed to generate the plot can be retrieved via
\verb{$data}. This is a \code{\link[tibble]{tibble}} containing the following
variables:
\itemize{
\item \code{V1}: the x-coordinate of each observation in the plane,
\item \code{V2}: the y-coordinate of each observation in the plane,
\item \code{Label}: the sample membership of each observation,
\item \code{Representation}: the type of matrix representation used to manipulate each
observation,
\item \code{Distance}: the distance used to measure how far each observation is from
the others.
}
}
\description{
This function generates 2-dimensional plots of samples of networks via
multi-dimensional scaling using all representations and distances included in
the package.
}
\details{
Match-Frobenius distance is the Frobenius distance considering networks in the Graph Space.
Match-Frobenius distance is computed via the graph matching algorithm with indefinite relaxation (via Frank-Wolfe), using \code{\link[iGraphMatch]{iGraphMatch}} \code{\link[iGraphMatch]{gm}} function (starting matrix is initialized to the barycenter and the maximum number of iterations to 20).
Match-Frobenius distance can be used only with adjacency matrix representation.
}
\examples{
gnp_params <- list(p = 1/3)
k_regular_params <- list(k = 8L)
x <- nvd(model = "gnp", n = 10L, model_params = gnp_params)
y <- nvd(model = "k_regular", n = 10L, model_params = k_regular_params)
ggplot2::autoplot(x, y)
plot(x, y)
}
