% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nvd-class.R
\name{nvd_from_adjacency_matrix}
\alias{nvd_from_adjacency_matrix}
\title{From adjacency matrices to \code{nvd} object}
\usage{
nvd_from_adjacency_matrix(adjmat, directed = FALSE, weighted = FALSE)
}
\arguments{
\item{adjmat}{A list of adjacency matrices.}

\item{directed}{A boolean specifying whether the function should consider the input matrices as directed or undirected. Default to \code{FALSE}.}

\item{weighted}{A boolean specifying whether the function should consider the input matrices as weighted or unweighted Default to \code{FALSE}.}
}
\value{
An \code{nvd} object.
}
\description{
This function creates an \code{nvd} object from a list of adjacency matrices.
}
\examples{
gnp_params <- list(p = 1/3)
x <- nvd(model = "gnp", n = 10L, model_params = gnp_params)
x_adj <- repr_nvd(x, representation = "adjacency")
nvd_from_adjacency_matrix(x_adj, directed = FALSE, weighted = FALSE)
}
