% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nvd-class.R
\name{nvd_num_vertices}
\alias{nvd_num_vertices}
\title{Summary of number of vertices in a sample of networks}
\usage{
nvd_num_vertices(x)
}
\arguments{
\item{x}{An \code{nvd} object.}
}
\value{
A \code{\link[base]{list}} with 5 components: a vector with the number of vertices of each network, the maximum, the minimum, the mean and the variance of the number of vertices of the networks.
}
\description{
This function summarizes information related to the number of vertices in an \code{nvd} object.
}
\examples{
gnp_params <- list(p = 1/3)
x <- nvd(model = "gnp", n = 10L, model_params = gnp_params)
nodes <- nvd_num_vertices(x)
nodes$num_vertices
nodes$max
}
