% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{plot.nvd}
\alias{plot.nvd}
\title{MDS Visualization of Network Distributions}
\usage{
\method{plot}{nvd}(x, y, ...)
}
\arguments{
\item{x}{A \code{\link{nvd}} object.}

\item{y}{A \code{\link{nvd}} object.}

\item{...}{Extra arguments to be passed to the plot function.}
}
\value{
Invisibly returns the dataset computed to generate the plot.
}
\description{
This function generates 2-dimensional plots of samples of networks via
multi-dimensional scaling using all represenations and distances included in
the package.
}
\examples{
x <- nvd("smallworld", 10)
y <- nvd("pa", 10)
plot(x, y)
}
