% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{power2}
\alias{power2}
\title{Power Simulations for Permutation Tests}
\usage{
power2(
  sample_size1,
  model1,
  params1,
  sample_size2,
  model2,
  params2,
  representation = "adjacency",
  distance = "frobenius",
  stats = c("flipr:t_ip", "flipr:f_ip"),
  B = 1000L,
  alpha = 0.05,
  test = "exact",
  k = 5L,
  R = 1000L
)
}
\arguments{
\item{sample_size1}{An integer specifying the size of the first sample.}

\item{model1}{A string specifying the model to be used for sampling networks
in the first sample. All \code{tidygraph::play_} functions are supported. The
model name corresponds to the name of the function without the \code{play_}
prefix.}

\item{params1}{A list specifying the parameters to be passed to the model
function that will generate the first sample.}

\item{sample_size2}{An integer specifying the size of the second sample.}

\item{model2}{A string specifying the model to be used for sampling networks
in the second sample. All \code{tidygraph::play_} functions are supported. The
model name corresponds to the name of the function without the \code{play_}
prefix.}

\item{params2}{A list specifying the parameters to be passed to the model}

\item{representation}{A string specifying the desired type of representation,
among: \code{"adjacency"}, \code{"laplacian"} and \code{"modularity"}.
Defaults to \code{"adjacency"}.}

\item{distance}{A string specifying the chosen distance for calculating the
test statistic, among: \code{"hamming"}, \code{"frobenius"},
\code{"spectral"} and \code{"root-euclidean"}. Defaults to
\code{"frobenius"}.}

\item{stats}{A character vector specifying the chosen test statistic(s),
among: \code{"original_edge_count"}, \code{"generalized_edge_count"},
\code{"weighted_edge_count"}, \code{"student_euclidean"}, \code{"welch_euclidean"} or any
statistics based on inter-point distances available in the \strong{flipr}
package: \code{"flipr:student_ip"}, \code{"flipr:fisher_ip"}, \code{"flipr:bg_ip"},
\code{"flipr:energy_ip"}, \code{"flipr:cq_ip"}. Defaults to \code{c("flipr:student_ip", "flipr:fisher_ip")}.}

\item{B}{The number of permutation or the tolerance. If this number is lower
than \code{1}, it is intended as a tolerance. Otherwise, it is intended as
the number of required permutations. Defaults to \code{1000L}.}

\item{alpha}{Significance level for hypothesis testing. Defaults to \code{0.05}.}

\item{test}{A character string specifying the formula to be used to compute
the permutation p-value. Choices are \code{"estimate"}, \code{"upper_bound"} and
\code{"exact"}. Defaults to \code{"exact"} which provides exact tests.}

\item{k}{An integer specifying the density of the minimum spanning tree used
for the edge count statistics. Defaults to \code{5L}.}

\item{R}{Number of Monte-Carlo trials used to estimate the power. Defaults to
\code{1000L}.}
}
\value{
A numeric value estimating the power of the test.
}
\description{
This function provides a Monte-Carlo estimate of the power of the permutation
tests proposed in this package.
}
\details{
Currently, six scenarios of pairs of populations are implemented. Scenario 0
allows to make sure that all our permutation tests are exact.
}
\examples{
gnp_params <- list(n = 24L, p = 1/3)
degree_params <- list(out_degree = rep(2, 24L), method = "configuration")
power2(
  sample_size1 = 10L, model1 = "gnp", params1 = gnp_params,
  sample_size2 = 10L, model2 = "degree", params2 = degree_params,
  R = 10L,
  B = 100L
)
}
