% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs.R
\name{power2_nbs}
\alias{power2_nbs}
\title{Power estimation of the NBS framework}
\usage{
power2_nbs(R = 100, n_obs = 10, p1 = 1/10, p2 = p1, n_nodes = 10,
  threshold = 0.95, B = 1000, alpha = 0.05, seed = NULL)
}
\arguments{
\item{R}{Number of simulation runs (default: \code{100L}).}

\item{n_obs}{Number of observations in the two samples (default: \code{10L}).}

\item{p1}{Probability of the presence of an edge in the first sample
(default: \code{0.1}).}

\item{p2}{Probability of the presence of an edge in the second sample
(default: \code{0.1}).}

\item{n_nodes}{Number of vertices in the network structure (default:
\code{10L}).}

\item{threshold}{Treshold to discard edges with low values of the test
statistic when searching for the connected components (default: \code{0.95}
which means 95th quantile of the null distribution of the test statistic).}

\item{B}{The number of permutation or the tolerance (default: \code{1000L}).
If this number is lower than \code{1}, it is intended as a tolerance.
Otherwise, it is intended as the number of required permutations.}

\item{alpha}{Significance level (default: \code{0.05}).}

\item{seed}{An integer for specifying the seed of the random generator for
result reproducibility (default: \code{NULL}).}
}
\value{
A matrix of size \code{(n_nodes x n_nodes)} reporting the estimated
  probability of rejection.
}
\description{
Power estimation of the NBS framework
}
\examples{
power2_nbs(R = 10L)
}
