% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{power_twosample}
\alias{power_twosample}
\title{Power Simulations for Permutation Tests}
\usage{
power_twosample(scenario = "0", n_pop = 4L, representation = "adjacency",
  distance = "frobenius", statistic = "lot", alpha = 0.05,
  test = "exact", k = 5L, R = 1000L, seed = NULL)
}
\arguments{
\item{scenario}{A character specifying the scenario to be simulated (choices
are: "0" [default], "A", "B", "C", "D" or "E").}

\item{n_pop}{The sample size of each population (the same is used for both
populations, default: 4L)}

\item{representation}{A string specifying the desired type of representation,
among: \code{"adjacency"} [default], \code{"laplacian"} and
\code{"modularity"}.}

\item{distance}{A string specifying the chosen distance for calculating the
test statistic, among: \code{"hamming"}, \code{"frobenius"} [default],
\code{"spectral"} and \code{"root-euclidean"}.}

\item{statistic}{A string specifying the chosen test statistic(s), among:
\code{"lot"} [default], \code{"sot"}, \code{"biswas"}, \code{"energy"},
\code{"student"}, \code{"welch"}, \code{"original"}, \code{"generalized"},
\code{"weighted"} or a combination from \code{c("lot", "sot", "biswas",
"energy")}.}

\item{alpha}{The significance level (default: \code{0.05}).}

\item{test}{A character string specifying if performing an exact test through
the use of Phipson-Smyth estimate of the p-value or an approximate test
through a Monte-Carlo estimate of the p-value (default: \code{"exact"}).}

\item{k}{An integer specifying the density of the minimum spanning tree used
for the edge count statistics (default: \code{5L}).}

\item{R}{The number of Monte-Carlo runs used to estimate the power (default:
1000L).}

\item{seed}{An integer specifying the seed to start randomness from (default:
uses clock).}
}
\value{
A vector \code{p} of \code{R} p-values, one for each Monte-Carlo run.
  The power can then be estimated as \code{mean(p <= alpha)}.
}
\description{
This function provides a Monte-Carlo estimate of the power of the permutation
tests proposed in this package.
}
\details{
Currently, six scenarios of pairs of populations are implemented. Scenario 0
allows to make sure that all our permutation tests are exact.
}
\examples{
alpha <- 0.05
p <- power_twosample(alpha = 0.05)
mean(p <= alpha)
}
