% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings.R
\name{push_to_graph_space}
\alias{push_to_graph_space}
\title{Graph Sample Embeddings}
\usage{
push_to_graph_space(obj)
}
\arguments{
\item{obj}{An object of class \code{\link{nvd}} containing the sample of graphs.}
}
\value{
An object of class \code{\link{nvd}} containing the sample of graphs in graph space.
}
\description{
A collection of functions to embed a sample of graphs into suitable spaces
for further statistical analysis.
}
\examples{
x <- nvd(sample_size = 5L, model = "gnp", n = 24L, p = 1/3)
x <- push_to_graph_space(x)
}
