% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representations.R
\name{repr_nvd}
\alias{repr_nvd}
\title{Network-Valued to Matrix-Valued Data}
\usage{
repr_nvd(x, y = NULL, representation = "laplacian")
}
\arguments{
\item{x}{An \code{\link{nvd}} object.}

\item{y}{An \code{\link{nvd}} object. If \code{NULL} (default), it is not
taken into account.}

\item{representation}{A string specifying the requested matrix
representation. Choices are: \code{"adjacency"}, \code{"laplacian"}
[default] or \code{"modularity"}.}
}
\value{
A list of matrices.
}
\description{
Network-Valued to Matrix-Valued Data
}
\examples{
x <- nvd("gnp", 10)
xm <- repr_nvd(x)
}
