% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs.R
\name{test2_nbs}
\alias{test2_nbs}
\title{Two-sample test of the NBS framework}
\usage{
test2_nbs(x, y, threshold = 0.95, B = 1000, seed = NULL)
}
\arguments{
\item{x}{An \code{\link{nvd}} object listing networks in sample 1.}

\item{y}{An \code{\link{nvd}} object listing networks in sample 2.}

\item{threshold}{Treshold to discard edges with low values of the test
statistic when searching for the connected components (default: \code{0.95}
which means 95th quantile of the null distribution of the test statistic).}

\item{B}{The number of permutation or the tolerance (default: \code{1000L}).
If this number is lower than \code{1}, it is intended as a tolerance.
Otherwise, it is intended as the number of required permutations.}

\item{seed}{An integer for specifying the seed of the random generator for
result reproducibility (default: \code{NULL}).}
}
\value{
A list containing information about the identified connected
  components and corresponding adjusted p-values.
}
\description{
Two-sample test of the NBS framework
}
\examples{
x <- replicate(10, igraph::sample_gnp(10, 1 / 10), simplify = FALSE)
y <- replicate(10, igraph::sample_gnp(10, 1 / 10), simplify = FALSE)
test2_nbs(x, y)
}
