% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{test_twosample}
\alias{test_twosample}
\title{Comparison of Network Distributions}
\usage{
test_twosample(x, y, representation = "adjacency", distance = "hamming",
  statistic = "mod", B = 1000L, alpha = 0.05, test = "exact",
  verbose = TRUE)
}
\arguments{
\item{x}{A \code{\link[base]{list}} of \code{\link[igraph]{igraph}} objects
or matrix representations of underlying networks from a given first
population.}

\item{y}{A \code{\link[base]{list}} of \code{\link[igraph]{igraph}} objects
or matrix representations of underlying networks from a given second
population.}

\item{representation}{A string specifying the desired type of representation,
among: \code{"adjacency"} [default], \code{"laplacian"} and
\code{"modularity"}.}

\item{distance}{A string specifying the chosen distance for calculating the
test statistic, among: \code{"hamming"} [default], \code{"frobenius"},
\code{"spectral"} and \code{"root-euclidean"}.}

\item{statistic}{A string specifying the chosen test statistic, among:
\code{"mod"} [default], \code{"dom"} or \code{"sdom"}.}

\item{B}{The number of permutation or the tolerance (default: \code{1000L}).
If this number is lower than \code{1}, it is intended as a tolerance.
Otherwise, it is intended as the number of required permutations.}

\item{alpha}{The significance level (default: \code{0.05}).}

\item{test}{A character string specifying if performing an exact test through
the use of Phipson Smyth p-value or an approximate test through a
Monte-Carlo p-value (default: \code{"exact"}).}

\item{verbose}{A boolean to switch on/off verbosity (information about
p-value resolution).}
}
\value{
A \code{\link[base]{list}} with three components: the value of the
  statistic for the original two samples, the p-value of the resulting
  permutation test and a numeric vector storing the values of the permutated
  statistics.
}
\description{
This function carries out an hypothesis test where the null hypothesis is
that the two populations of networks share the same underlying probabilistic
distribution against the alternative hypothesis that the two populations come
from different distributions. The test is performed in a non-parametric
fashion using a permutational framework in which several statistics can be
used, together with several choices of network matrix representations and
distances between networks.
}
\details{
The exhaustive list of all possible permutations is used if its length is
smaller than either the user-defined number of random permutations requested
or the number of permutations required for achieving a user-defined p-value
resolution. Otherwise, the p-value is estimated via Monte-Carlo simulations.
}
\examples{
n <- 10L

# Two different models for the two populations
x <- nvd("smallworld", n)
y <- nvd("pa", n)
test1 <- test_twosample(x, y, "modularity")
test1$pvalue

# Same model for the two populations
x <- nvd("smallworld", n)
y <- nvd("smallworld", n)
test2 <- test_twosample(x, y, "modularity")
test2$pvalue
}
