% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nvd_class.R
\name{var_nvd}
\alias{var_nvd}
\title{Fréchet Variance of Network-Valued Data}
\usage{
var_nvd(x, x0, distance = "frobenius")
}
\arguments{
\item{x}{An \code{\link{nvd}} object listing a sample of networks.}

\item{x0}{A network already in matrix representation around which to
calculate variance (usually the Fréchet mean but not necessarily). Note
that the chosen matrix representation is extracted from this parameter.}

\item{distance}{A string specifying the distance to be used. Possible choices
are: hamming, frobenius [default], spectral or root-euclidean. When the
Fréchet mean is used as \code{x0} parameter, the distance should match the
one used to compute the mean. This is not currently checked.}
}
\value{
A positive scalar value evaluating the amount of variability of the
  sample around the specified network.
}
\description{
Fréchet Variance of Network-Valued Data
}
\examples{
d <- nvd(n = 10L)
m <- mean(d)
var_nvd(x = d, x0 = m, distance = "frobenius")
}
