% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nvd-class.R
\name{as_nvd}
\alias{as_nvd}
\title{Network-Valued Data Class}
\usage{
as_nvd(obj, matched = FALSE, verbose = FALSE)
}
\arguments{
\item{obj}{A list of R objects convertible into
\code{\link[tidygraph]{tbl_graph}} objects.}

\item{matched}{A boolean specifying whether graphs should be considered
node-matched or not. Defaults to \code{FALSE}.}

\item{verbose}{A boolean specifying whether to print some information about
the sample. Defaults to \code{FALSE}.}
}
\value{
An \code{\link{nvd}} object.
}
\description{
This function flags a list of \code{\link[igraph]{igraph}} objects as an
\code{\link{nvd}} object as defined in this package.
}
\examples{
as_nvd(nvd("smallworld", 10))
}
