% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nvd_class.R
\name{as_nvd}
\alias{as_nvd}
\title{Coercion to Network-Valued Data Object}
\usage{
as_nvd(obj)
}
\arguments{
\item{obj}{A list of \code{\link[igraph]{igraph}} objects.}
}
\value{
An \code{\link{nvd}} object.
}
\description{
This function flags a list of \code{\link[igraph]{igraph}} objects as an
\code{\link{nvd}} object as defined in this package.
}
\examples{
as_nvd(nvd("smallworld", 10))
}
