% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nvd_class.R
\name{mean.nvd}
\alias{mean.nvd}
\title{Fréchet Mean of Network-Valued Data}
\usage{
\method{mean}{nvd}(x, representation = "adjacency", ...)
}
\arguments{
\item{x}{An \code{\link{nvd}} object.}

\item{representation}{A string specifying the graph representation to be used (choices: adjacency [default], laplacian, modularity).}

\item{...}{Other argument to be parsed to the \code{\link[base]{mean}} function.}
}
\value{
The mean network in the chosen matrix representation assuming Euclidean geometry for now.
}
\description{
Fréchet Mean of Network-Valued Data
}
\examples{
d <- nvd(n = 10L)
mean(d)
}
