% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_test.R
\name{network_test2p}
\alias{network_test2p}
\title{Comparison of Network Distributions}
\usage{
network_test2p(x, y, representation = "adjacency", distance = "hamming",
  statistic = "average", B = 1000L, alpha = 0.05)
}
\arguments{
\item{x}{A \code{\link[base]{list}} of \code{\link[igraph]{igraph}} objects
or matrix representations of underlying networks from a given first
population.}

\item{y}{A \code{\link[base]{list}} of \code{\link[igraph]{igraph}} objects
or matrix representations of underlying networks from a given second
population.}

\item{representation}{A string specifying the desired type of representation,
among: \code{"adjacency"} [default], \code{"laplacian"} and
\code{"modularity"}.}

\item{distance}{A string specifying the chosen distance for calculating the
test statistic, among: \code{"hamming"} [default], \code{"frobenius"},
\code{"spectral"} and \code{"root-euclidean"}.}

\item{statistic}{A string specifying the chosen test statistic, among:
\code{"average"} [default] or \code{"frechet"}.}

\item{B}{The number of permutation or the tolerance (default: \code{1000L}).
If this number is lower than \code{1}, it is intended as a tolerance.
Otherwise, it is intended as the number of required permutations.}

\item{alpha}{The significance level (default: \code{0.05}).}
}
\value{
A \code{\link[base]{list}} with two components: the value of the
  statistic for the original two samples and the p-value of the resulting
  permutation test.
}
\description{
This function carries out an hypothesis test where the null hypothesis is
that the two populations of networks share the same underlying probabilistic
distribution against the alternative hypothesis that the two populations come
from different distributions. The test is performed in a non-parametric
fashion using a permutational framework in which several statistics can be
used, together with several choices of network matrix representations and
distances between networks.
}
\details{
The exhaustive list of all possible permutations is used if its length is
smaller than either the user-defined number of random permutations requested
or the number of permutations required for achieving a user-defined p-value
resolution. Otherwise, the p-value is estimated via Monte-Carlo simulations.
}
\examples{
n <- 25L

x <- list()
y <- list()
for (i in 1:10) {
  x[[i]] <- igraph::watts.strogatz.game(1, n, 3, 0.05)
  y[[i]] <- igraph::barabasi.game(n, m = 3, power = 2, directed = FALSE)
}
test1 <- network_test2p(x, y, "modularity")
test1

x <- list()
y <- list()
for (i in 1:10) {
  x[[i]] <- igraph::watts.strogatz.game(1, n, 3, 0.05)
  y[[i]] <- igraph::watts.strogatz.game(1, n, 3, 0.05)
}

test2 <- network_test2p(x, y, "modularity")
test2
}
