% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nevada-package.R
\docType{package}
\name{nevada-package}
\alias{nevada}
\alias{nevada-package}
\title{nevada: NEtwork-VAlued Data Analysis}
\description{
nevada provides a ...
}
\details{
It has three main goals:

\itemize{
\item ...
\item ...
\item ...
}

To learn more about nevada, start with the vignettes:
\code{browseVignettes(package = "nevada")}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://astamm.github.io/nevada}
  \item \url{https://github.com/astamm/nevada}
  \item Report bugs at \url{https://github.com/astamm/nevada/issues}
}

}
\author{
\strong{Maintainer}: Aymeric Stamm \email{aymeric.stamm@math.cnrs.fr} (\href{https://orcid.org/0000-0002-8725-3654}{ORCID})

Authors:
\itemize{
  \item Ilenia Lovato \email{ilenia.lovato01@universitadipavia.it}
  \item Alessia Pini \email{alessia.pini@unicatt.it}
  \item Simone Vantini \email{simone.vantini@polimi.it}
}

}
