% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nvd_class.R
\name{nvd}
\alias{nvd}
\title{Network-Valued Data Constructor}
\usage{
nvd(model = "smallworld", n = 0L, pref.matrix = NULL, lambda = NULL)
}
\arguments{
\item{model}{A string specifying the model to be used for sampling networks
(current choices are: \code{"sbm"}, \code{"k_regular"}, \code{"gnp"},
\code{"smallworld"} [default], \code{"pa"}, \code{"poisson"}).}

\item{n}{An integer specifying the sample size (default: \code{0L}).}

\item{pref.matrix}{A matrix giving the Bernoulli rates for the SBM
generator (see \code{\link[igraph]{sample_sbm}} for details).}

\item{lambda}{A numeric value specifying the mean value for the Poisson
generator.}
}
\value{
A \code{nvd} object which is a list of \code{\link[igraph]{igraph}}
  objects.
}
\description{
This is the constructor for objects of class \code{nvd}.
}
\examples{
x <- nvd(n = 10L)
x
}
