% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nvd-class.R
\name{nvd}
\alias{nvd}
\title{Network-Valued Data Constructor}
\usage{
nvd(sample_size, model, ...)
}
\arguments{
\item{sample_size}{An integer specifying the sample size.}

\item{model}{A string specifying the model to be used for sampling networks.
All \code{tidygraph::play_} functions are supported. The model name corresponds
to the name of the function without the \code{play_} prefix.}

\item{...}{Model parameters to be passed to the model function.}
}
\value{
A \code{\link{nvd}} object which is a list of \code{\link[tidygraph:tbl_graph]{tidygraph::tbl_graph}} objects.
}
\description{
This is the constructor for objects of class \code{nvd}.
}
\examples{
params <- list(n_dim = 1L, dim_size = 4L, order = 4L, p_rewire = 0.15)
nvd(sample_size = 1L, model = "smallworld", !!!params)
}
