% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representations.R
\name{repr_nvd}
\alias{repr_nvd}
\title{Network-Valued to Matrix-Valued Data}
\usage{
repr_nvd(x, y = NULL, representation = "adjacency")
}
\arguments{
\item{x}{An \code{\link{nvd}} object.}

\item{y}{An \code{\link{nvd}} object. If \code{NULL} (default), it is not
taken into account.}

\item{representation}{A string specifying the requested matrix
representation. Choices are: \code{"adjacency"}, \code{"laplacian"} or
\code{"modularity"}. Default is \code{"adjacency"}.}
}
\value{
A list of matrices.
}
\description{
Network-Valued to Matrix-Valued Data
}
\examples{
gnp_params <- list(p = 1/3)
x <- nvd(model = "gnp", n = 10L, model_params = gnp_params)
xm <- repr_nvd(x)
}
