% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplers.R
\name{samplers}
\alias{samplers}
\alias{play_poisson}
\alias{rpois_network}
\alias{play_exponential}
\alias{rexp_network}
\alias{play_binomial}
\alias{rbinom_network}
\title{Graph samplers using edge distributions}
\usage{
play_poisson(num_vertices, lambda = 1, directed = FALSE, loops = FALSE)

rpois_network(n, num_vertices, lambda = 1, directed = FALSE, loops = FALSE)

play_exponential(num_vertices, rate = 1, directed = FALSE, loops = FALSE)

rexp_network(n, num_vertices, rate = 1, directed = FALSE, loops = FALSE)

play_binomial(
  num_vertices,
  size = 1,
  prob = 0.5,
  directed = FALSE,
  loops = FALSE
)

rbinom_network(
  n,
  num_vertices,
  size = 1,
  prob = 0.5,
  directed = FALSE,
  loops = FALSE
)
}
\arguments{
\item{num_vertices}{Number of vertices.}

\item{lambda}{The mean parameter for the Poisson distribution (default: 1).}

\item{directed}{A boolean specifying whether to generate directed or undirected graphs. Defaults to \code{FALSE}.}

\item{loops}{A boolean specifying whether loops are authorized. Defaults to \code{FALSE}.}

\item{n}{Sample size.}

\item{rate}{The rate parameter for the exponential distribution (default: 1).}

\item{size}{The number of trials for the binomial distribution (default: 1).}

\item{prob}{The probability of success on each trial for the binomial
distribution (default: 0.5).}
}
\value{
A object of class \code{\link{nvd}} containing the sample of graphs.
}
\description{
A collection of functions to generate random graphs with specified edge
distribution.
}
\examples{
nvd <- rexp_network(10, 68)
}
