% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{statistics}
\alias{statistics}
\alias{stat_mod}
\alias{stat_dom}
\alias{stat_sdom}
\title{Test Statistics for Network Populations}
\usage{
stat_mod(d, indices)

stat_dom(d, indices)

stat_sdom(d, indices)
}
\arguments{
\item{d}{A matrix of dimension \eqn{(n1+n2)x(n1+n2)} containing the distances
between all the elements of the two samples put together.}

\item{indices}{A vector of dimension \eqn{n1} containing the indices of the
elements of the first sample.}
}
\value{
A scalar giving the value of the desired test statistic.
}
\description{
This is a collection of functions that provide various test statistics for
populations of networks. The \emph{MoD} (mean of the distances) test
statistic is given by the mean of all the distances between the \eqn{n1}
elements of the fist sample and those (\eqn{n2}) of the second one. The
\emph{DoM} (distance of the means) test statistic is given by the distance
between the Frechet mean of the first sample (with cardinality \eqn{n1}) and
that of the second one (with cardinality \eqn{n2}). The \emph{sDoM}
(standardized distance of the means) is the \emph{DoM} statistic standardized
by the pooled sample standard deviation. The distance used to compute the
value of the test statistic is the same used to find the Fréchet means.
}
\examples{
x <- nvd("smallworld", 10)
y <- nvd("pa", 10)
d <- dist_nvd(x, y, "adjacency", "spectral")
stat_mod(d, 1:10)
stat_dom(d, 1:10)
stat_sdom(d, 1:10)
}
