% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{statistics}
\alias{statistics}
\alias{stat_lot}
\alias{stat_sot}
\alias{stat_biswas}
\alias{stat_energy}
\alias{stat_mod}
\alias{stat_dom}
\alias{stat_student_euclidean}
\alias{stat_welch_euclidean}
\alias{stat_edge_count}
\title{Test Statistics for Network Populations}
\usage{
stat_lot(d, indices)

stat_sot(d, indices)

stat_biswas(d, indices)

stat_energy(d, indices, alpha = 1)

stat_mod(d, indices)

stat_dom(d, indices, standardize = TRUE)

stat_student_euclidean(d, indices)

stat_welch_euclidean(d, indices)

stat_edge_count(d, indices, type = "original")
}
\arguments{
\item{d}{Either a matrix of dimension \eqn{(n1+n2)x(n1+n2)} containing the
distances between all the elements of the two samples put together or, for
the \code{stat_dom_frobenius} statistic, the concatenation of the lists of
matrix representations of networks in samples 1 and 2.}

\item{indices}{A vector of dimension \eqn{n1} containing the indices of the
elements of the first sample.}

\item{standardize}{If \code{TRUE} (default), standardize the DoM statistic
using the pooled sample variance estimator.}

\item{type}{A string specifying the version of the edge count test statistic
to be used. Choices are \code{"original"} [default], \code{"generalized"}
or \code{"weighted"}.}
}
\value{
A scalar giving the value of the desired test statistic.
}
\description{
This is a collection of functions that provide various test statistics for
populations of networks. The \emph{MoD} (mean of the distances) test
statistic is given by the mean of all the distances between the \eqn{n1}
elements of the fist sample and those (\eqn{n2}) of the second one. The
\emph{DoM} (distance of the means) test statistic is given by the distance
between the Frechet mean of the first sample (with cardinality \eqn{n1}) and
that of the second one (with cardinality \eqn{n2}). The \emph{sDoM}
(standardized distance of the means) is the \emph{DoM} statistic standardized
by the pooled sample standard deviation. The distance used to compute the
value of the test statistic is the same used to find the Fréchet means.
}
\examples{
n1 <- 30L
n2 <- 10L
x <- nvd("smallworld", n1)
y <- nvd("pa", n2)
d <- dist_nvd(x, y, representation = "laplacian", distance = "frobenius")
stat_lot(d, 1:n1)
stat_sot(d, 1:n1)
stat_biswas(d, 1:n1)
stat_energy(d, 1:n1)
r <- repr_nvd(x, y, representation = "laplacian")
stat_student_euclidean(r, 1:n1)
stat_welch_euclidean(r, 1:n1)
e <- edge_count_global_variables(d, n1, k = 5L)
stat_edge_count(e, 1:n1, type = "original")
stat_edge_count(e, 1:n1, type = "generalized")
stat_edge_count(e, 1:n1, type = "weighted")
}
