% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{test2_global}
\alias{test2_global}
\title{Global Two-Sample Test for Network-Valued Data}
\usage{
test2_global(
  x,
  y,
  representation = "adjacency",
  distance = "frobenius",
  statistic = "lot",
  B = 1000L,
  test = "exact",
  k = 5L,
  seed = NULL
)
}
\arguments{
\item{x}{An \code{\link{nvd}} object listing networks in sample 1.}

\item{y}{An \code{\link{nvd}} object listing networks in sample 2.}

\item{representation}{A string specifying the desired type of representation,
among: \code{"adjacency"} [default], \code{"laplacian"} and
\code{"modularity"}.}

\item{distance}{A string specifying the chosen distance for calculating the
test statistic, among: \code{"hamming"}, \code{"frobenius"} [default],
\code{"spectral"} and \code{"root-euclidean"}.}

\item{statistic}{A string specifying the chosen test statistic(s), among:
\code{"lot"} [default], \code{"sot"}, \code{"biswas"}, \code{"energy"},
\code{"student"}, \code{"welch"}, \code{"original"}, \code{"generalized"},
\code{"weighted"} or a combination from \code{c("lot", "sot", "biswas",
"energy")}.}

\item{B}{The number of permutation or the tolerance (default: \code{1000L}).
If this number is lower than \code{1}, it is intended as a tolerance.
Otherwise, it is intended as the number of required permutations.}

\item{test}{A character string specifying if performing an exact test through
the use of Phipson-Smyth estimate of the p-value or an approximate test
through a Monte-Carlo estimate of the p-value (default: \code{"exact"}).}

\item{k}{An integer specifying the density of the minimum spanning tree used
for the edge count statistics (default: \code{5L}).}

\item{seed}{An integer for specifying the seed of the random generator for
result reproducibility (default: \code{NULL}).}
}
\value{
A \code{\link[base]{list}} with three components: the value of the
  statistic for the original two samples, the p-value of the resulting
  permutation test and a numeric vector storing the values of the permuted
  statistics.
}
\description{
This function carries out an hypothesis test where the null hypothesis is
that the two populations of networks share the same underlying probabilistic
distribution against the alternative hypothesis that the two populations come
from different distributions. The test is performed in a non-parametric
fashion using a permutational framework in which several statistics can be
used, together with several choices of network matrix representations and
distances between networks.
}
\examples{
n <- 10L

# Two different models for the two populations
x <- nvd("smallworld", n)
y <- nvd("pa", n)
t1 <- test2_global(x, y, "modularity")
t1$pvalue

# Same model for the two populations
x <- nvd("smallworld", n)
y <- nvd("smallworld", n)
t2 <- test2_global(x, y, "modularity")
t2$pvalue
}
