% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{test2_local}
\alias{test2_local}
\title{Local Two-Sample Test for Network-Valued Data}
\usage{
test2_local(x, y, partition, representation = "adjacency",
  distance = "frobenius", statistic = "lot", B = 1000L,
  alpha = 0.05, test = "exact", k = 5L, seed = NULL)
}
\arguments{
\item{x}{An \code{\link{nvd}} object listing networks in sample 1.}

\item{y}{An \code{\link{nvd}} object listing networks in sample 2.}

\item{partition}{Either a list or a vector specifying vertex memberships into
partition elements.}

\item{representation}{A string specifying the desired type of representation,
among: \code{"adjacency"} [default], \code{"laplacian"} and
\code{"modularity"}.}

\item{distance}{A string specifying the chosen distance for calculating the
test statistic, among: \code{"hamming"}, \code{"frobenius"} [default],
\code{"spectral"} and \code{"root-euclidean"}.}

\item{statistic}{A string specifying the chosen test statistic(s), among:
\code{"lot"} [default], \code{"sot"}, \code{"biswas"}, \code{"energy"},
\code{"student"}, \code{"welch"}, \code{"original"}, \code{"generalized"},
\code{"weighted"} or a combination from \code{c("lot", "sot", "biswas",
"energy")}.}

\item{B}{The number of permutation or the tolerance (default: \code{1000L}).
If this number is lower than \code{1}, it is intended as a tolerance.
Otherwise, it is intended as the number of required permutations.}

\item{alpha}{Significance level for hypothesis testing (default:
\code{0.05}). If set to 1, the function outputs properly adjusted p-values.
If lower than 1, then only p-values lower than alpha are properly adjusted.}

\item{test}{A character string specifying if performing an exact test through
the use of Phipson-Smyth estimate of the p-value or an approximate test
through a Monte-Carlo estimate of the p-value (default: \code{"exact"}).}

\item{k}{An integer specifying the density of the minimum spanning tree used
for the edge count statistics (default: \code{5L}).}

\item{seed}{An integer for specifying the seed of the random generator for
result reproducibility (default: \code{NULL}).}
}
\value{
A length-2 list reporting the adjusted p-values of each element of
  the partition for the intra- and inter-tests.
}
\description{
Local Two-Sample Test for Network-Valued Data
}
\examples{
n <- 10
p1 <- matrix(
  data = c(0.1, 0.4, 0.1, 0.4,
           0.4, 0.4, 0.1, 0.4,
           0.1, 0.1, 0.4, 0.4,
           0.4, 0.4, 0.4, 0.4),
  nrow = 4,
  ncol = 4,
  byrow = TRUE
)
p2 <- matrix(
  data = c(0.1, 0.4, 0.4, 0.4,
           0.4, 0.4, 0.4, 0.4,
           0.4, 0.4, 0.1, 0.1,
           0.4, 0.4, 0.1, 0.4),
  nrow = 4,
  ncol = 4,
  byrow = TRUE
)
set.seed(1234)
sim <- n \%>\%
  purrr::rerun(
    x = igraph::sample_sbm(68, p1, c(17,17,17,17)),
    y = igraph::sample_sbm(68, p2, c(17,17,17,17))
  ) \%>\%
  purrr::transpose() \%>\%
  purrr::map(as_nvd)
t0 <- test2_local(sim$x, sim$y, m, statistic = c("lot", "sot"), seed = 1234)
}
